[NFC] Refactor out the dropped-block optimization code in MergeBlocks (#6982)
This just moves the code out into a function. A later PR will use it in another place.

Add a test that shows the motivation for that later PR: we fail to optimize away
a block return value at the top level of a function. Fixing that will involve calling
the new function here from another place.