Optimize `Module::get_*` family of functions with `std::string_view` in `Module::getModuleElement`
Passing a constant string to functions requires memory allocation, and allocation is inherently slow.
Since we are using C++17, we can use `string_view` and remove this unnecessary allocation.

Although the code seems simple enough for the optimizer to remove this allocation after inlining, tests on Clang 18 show that this is not the case.