Don't use multiple tables in wasm-split
Currently, wasm-split will create a second table rather than replacing
the existing one when reftypes is supported (or rather whether a
target_features section is present indicating it's already used).
But the emscripten integration code doesn't currently support this,
and it's not tested anywhere. The existing tests started failing when
Binaryen stopped stripping the features section by default.
(But some users would already be broken by this). For now, removing
this untested behavior is a safe default.