Make 32-bit hashing identical to 64-bit in TypeSSA (#7048)
This is NFC on 64-bit systems but noticeable on 32.

Also remove the 32-bit path in hash_combine. That isn't necessary for this fix,
but it makes the code simpler and also makes debugging between systems
simpler. It might also avoid problems in future cases, if we are lucky. The only
cost is perhaps a slight slowdown on 32-bit systems, which seems worth it.

Fixes #7046