Read the names section first
Rather than back-patching names when we get to the names section in the binary
reader, skip ahead to read the names section before anything else so we can use
the final names right away. This is a prerequisite for using IRBuilder in the
binary reader.

The only functional change is that we now allow empty local names. Empty names
are perfectly valid.