Use empty blocks instead of nops for empty scopes in IRBuilder
When IRBuilder builds an empty non-block scope such as a function body,
an if arm, a try block, etc, it needs to produce some expression to
represent the empty contents. Previously it produced a nop, but change
it to produce an empty block instead. The binary writer and printer have
special logic to elide empty blocks, so this produces smaller output.

Update J2CLOpts to recognize functions containing empty blocks as
trivial to avoid regressing one of its tests.