Note subtyping for tuple instructions in Unsubtyping
Unsubtyping uses the subtype-exprs.h utility to collect subtyping
constraints from the module, but that utility did not properly note
subtyping in tuple instructions. Normally this is not a problem because
the tuple.make operand types exactly match the elements of the result
tuple type, but it is also possible for the operands to be more refined,
in which case this missing logic can lead to bugs.

Fixes #7087.