Special-case unreachable If conditions in IRBuilder
When IRBuilder completes a scope that contains an unreachable
expression, it has to find the last unreachable expression and drop any
concretely typed expression that precedes it. If the original
unreachable expression has been made a child of some other expression,
that's usually fine because that other expression will become
unreachable as well. However, if conditions are special because they can
be unreachable without making their ifs unreachable, so we have to work
harder to find unreachable if conditions.