[NFC] Refactor nice methods in fuzz_shell.js (#7096)
This mostly moves the code around and avoids some duplication. It
also tracks the list of exports with both names and values, so that
if we compile more than one module, we can still access exports
from the previous.

Also add a first test of running fuzz_shell.js in node.

This does make build() append the exports, which was done before
on the main module but not the second one. That only affects the
wasm-split fuzzer, which is not active yet, so this is still NFC.