Fixup block-nested pops even when EH is not enabled
While parsing a binary file, there may be pops that need to be fixed up
even if EH is not (yet) enabled because the target features section has
not been parsed yet. Previously `EHUtils::handleBlockNestedPops` did not
do anything if EH was not enabled, so the binary parser would fail to
fix up pops in that case. Fix the utility to run no matter what features
are enabled and fix up its users so it is only called from optimization
passes when EH is enabled.

Fixes #7127.